+++ src/lib/core/profileManager.h
@@ -44,7 +44,7 @@ const int          DEFAULT_MAX_PARTIAL_PREDICTION_SIZE = 100;
 const std::string  DEFAULT_COMBINATION_POLICY          = "Meritocracy";
 const std::string  DEFAULT_PREDICTIVE_PLUGINS          = "";
 const int          DEFAULT_MAX_BUFFER_SIZE             = 1024;  // TODO: <==== not implemented yet
-const int          DEFAULT_SUGGESTIONS                 = 6;
+const size_t       DEFAULT_SUGGESTIONS                 = 6;
 const bool         DEFAULT_REPEAT_SUGGESTION           = false;
 const unsigned int DEFAULT_GREEDY_SUGGESTION_THRESHOLD = 0;
 const std::string  DEFAULT_PLUGINS                     = "";
