+++ test/lib/plugins/newSmoothedNgramPluginTest.cpp
@@ -79,32 +79,32 @@ void NewSmoothedNgramPluginTest::testLearning()
     {
 	ct->update("f");
 	Prediction actual = plugin->predict(SIZE);
-	CPPUNIT_ASSERT_EQUAL(0, actual.size());
+	CPPUNIT_ASSERT_EQUAL(static_cast<size_t>(0), actual.size());
     }
 
     {
 	ct->update("o");
 	Prediction actual = plugin->predict(SIZE);
-	CPPUNIT_ASSERT_EQUAL(0, actual.size());
+	CPPUNIT_ASSERT_EQUAL(static_cast<size_t>(0), actual.size());
     }
 
     {
 	ct->update("o ");
 	Prediction actual = plugin->predict(SIZE);
-	CPPUNIT_ASSERT_EQUAL(1, actual.size());
+	CPPUNIT_ASSERT_EQUAL(static_cast<size_t>(1), actual.size());
 	CPPUNIT_ASSERT_EQUAL(std::string("foo"), actual.getSuggestion(0).getWord());
     }
 
     {
 	ct->update("bar");
 	Prediction actual = plugin->predict(SIZE);
-	CPPUNIT_ASSERT_EQUAL(0, actual.size());
+	CPPUNIT_ASSERT_EQUAL(static_cast<size_t>(0), actual.size());
     }
 
     {
 	ct->update(" ");
 	Prediction actual = plugin->predict(SIZE);
-	CPPUNIT_ASSERT_EQUAL(2, actual.size());
+	CPPUNIT_ASSERT_EQUAL(static_cast<size_t>(2), actual.size());
 	CPPUNIT_ASSERT_EQUAL(std::string("foo"), actual.getSuggestion(0).getWord());
 	CPPUNIT_ASSERT_EQUAL(std::string("bar"), actual.getSuggestion(1).getWord());
     }
@@ -112,7 +112,7 @@ void NewSmoothedNgramPluginTest::testLearning()
     {
 	ct->update("foobar ");
 	Prediction actual = plugin->predict(SIZE);
-	CPPUNIT_ASSERT_EQUAL(3, actual.size());
+	CPPUNIT_ASSERT_EQUAL(static_cast<size_t>(3), actual.size());
 	CPPUNIT_ASSERT_EQUAL(std::string("foobar"), actual.getSuggestion(0).getWord());
 	CPPUNIT_ASSERT_EQUAL(std::string("foo"), actual.getSuggestion(1).getWord());
 	CPPUNIT_ASSERT_EQUAL(std::string("bar"), actual.getSuggestion(2).getWord());
@@ -121,7 +121,7 @@ void NewSmoothedNgramPluginTest::testLearning()
     {
 	ct->update("f");
 	Prediction actual = plugin->predict(SIZE);
-	CPPUNIT_ASSERT_EQUAL(2, actual.size());
+	CPPUNIT_ASSERT_EQUAL(static_cast<size_t>(2), actual.size());
 	CPPUNIT_ASSERT_EQUAL(std::string("foobar"), actual.getSuggestion(0).getWord());
 	CPPUNIT_ASSERT_EQUAL(std::string("foo"), actual.getSuggestion(1).getWord());
     }
